+++ app/main.py
@@ -3,9 +3,10 @@ from os import environ
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
+from app.routers import esm3_model
 from middleware import ExceptionHandler, TokenAuthentication, ResponseTime
 
-from routers import esm3_generate, health
+from routers import health
 
 from model import ESM3Model
 
@@ -18,19 +19,18 @@ hf_token = environ.get("HF_TOKEN", "")
 api_token = environ.get("API_TOKEN", "")
 model_name = environ.get("MODEL_NAME", "esm3-open")
 quantize = environ.get("QUANTIZE", "false").lower() == "true"
-quant_group_size = int(environ.get("QUANT_GROUP_SIZE", 8))
 device = environ.get("DEVICE", "cpu")
 
 app = FastAPI(
     title="ESM3 Inference Server",
     description="ESM3 evolutionary protein modelling inference server.",
-    version="0.0.8",
+    version="0.0.9",
 )
 
 # The ESM3 model requires a license agreement.
 hf_login(token=hf_token)
 
-model = ESM3Model(model_name, quantize, quant_group_size, device)
+model = ESM3Model(model_name, quantize, device)
 
 app.state.model = model
 
@@ -49,7 +49,7 @@ if api_token:
 
 app.add_middleware(ResponseTime)
 
-app.include_router(esm3_generate.router)
+app.include_router(esm3_model.router)
 app.include_router(health.router)