+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.9
+
+- Removed group quantization.
+- Added model info endpoint.
+
 # 0.0.8
 
 - Added quantization group size setting.